requirements.txt
alabaster==0.7.16
asgiref==3.7.2
Babel==2.15.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cmake==3.28.1
colorama==0.4.6
cryptography==42.0.2
Cython==3.0.7
Django==5.0
django-appconf==1.0.6
django-browser-reload==1.12.1
django-compressor==4.4
django-cors-headers==4.3.1
django-filter==24.2
django-otp==1.3.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
firebase-admin==6.4.0
google-api-core==2.16.1
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
hedera-sdk-py==2.29.0
httplib2==0.22.0
idna==3.6
imagesize==1.4.1
install-jdk==1.1.0
jdk4py==17.0.9.2
Jinja2==3.1.4
Markdown==3.6
MarkupSafe==2.1.5
msgpack==1.0.7
numpy==1.26.3
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.2
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.18.0
pyjnius==1.6.1
PyJWT==2.8.0
pyparsing==3.1.1
pypng==0.20220715.0
python-nmap==0.7.1
python-screen==1.0.3
pywatchman==1.4.1
pyzbar==0.1.9
qrcode==7.4.2
rcssmin==1.1.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
rjsmin==1.2.1
rsa==4.9
setuptools==69.0.2
snowballstemmer==2.2.0
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.4.4
typing==3.7.4.3
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.0
wheel==0.42.0
